+++ .gitignore
@@ -1,3 +1,4 @@
+docs
 lib
 build
 node_modules
