--- documentation.yaml
@@ -1,2 +1,0 @@
-toc:
-  - ViberPlay
