+++ src/models/ad-instance.ts
