+++ src/models/signed-player-info.ts
